+++ lib/tools/tools.dart
@@ -1,3 +1,5 @@
 export 'changelog/changelog_command.dart';
 export 'doctor/doctor_command.dart';
 export 'init/init_command.dart';
+export 'rename/rename_command.dart';
+export 'upgrade/upgrade_command.dart';